+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## v0.0.1 - 202?-??-?? - ??
+## TODO: v0.0.1 - 2022-01-03 - Moving
 
 #### Nothworthy Changes
 
-...
+* Initial extraction of Route53Provider from octoDNS core
+* Initial extraction of AwsAcmMangingProcessor
 
 #### Stuff
 
-...
+* Nothing